Moved all classes to gsi_ncdiag.py; need to modify scripts that call the classes still
