+++ src/gsi-ncdiag/CMakeLists.txt
@@ -1,11 +1,5 @@
 list( APPEND libs
-  conv.py
-  radiances.py
-  composition.py
-  conv_dicts.py
-  rad_dicts.py
-  aod_dicts.py
-  oz_dicts.py
+  gsi_ncdiag.py
 )
 
 list( APPEND programs
